+++ common/src/main/java/de/cech12/unlitcampfire/ModTags.java
@@ -3,8 +3,7 @@ package de.cech12.unlitcampfire;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.tags.TagKey;
 import net.minecraft.world.item.Item;
-
-import javax.annotation.Nonnull;
+import org.jetbrains.annotations.NotNull;
 
 public class ModTags {
 
@@ -12,7 +11,7 @@ public class ModTags {
 
         public static final TagKey<Item> MAKES_CAMPFIRE_INFINITE = tag("makes_campfire_infinite");
 
-        private static TagKey<Item> tag(@Nonnull String name) {
+        private static TagKey<Item> tag(@NotNull String name) {
             return TagKey.create(Registries.ITEM, Constants.id(name));
         }
     }