+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [this versioning scheme](https://gist.github.com/cech12/69319028e88c50349a6b044000a6607b).
 
+## [1.21.11-3.6.0.0] - 2025-12-14
+### Changed
+- updated to Minecraft 1.21.11 (Fabric 0.139.5+1.21.11, NeoForge 21.11.6-beta, Forge 61.0.2)
+- updated Cloth Config support (21.11.151) (Fabric/Quilt)
+- updated ModMenu support (17.0.0-alpha.1) (Fabric/Quilt)
+- updated Jade support (21.0.1) (Fabric/Quilt/NeoForge)
+
 ## [1.21.9-3.5.0.1] - 2025-11-11
 ### Fixed
 - config overrides of Lithium was not correct in NeoForge